requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
boto==2.49.0
boto3==1.14.56
botocore==1.17.56
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
Django==3.1
django-cors-headers==3.5.0
django-storages==1.10
djangorestframework==3.11.1
docutils==0.15.2
idna==2.10
image==1.5.33
isort==5.4.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pillow==8.0.1
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
